Handle locations with more than 30 chars for lat/lon.
For some locations more than 30 characters are needed for
representing the lat/lon information and the null character
at the end of the string is overwritten.
This causes the next member of the struct to be included when
reading the string which causes a NumberFormatException in
convertRationalLatLonToFloat() in ExifInterface.java and a
0f is returned.

The size of GpsLongRaw/GpsLatRaw is increased to 72 in
jhead.h since this is the maximum length of the string,
i.e. 11 * 6 + 3(‘/’) + 2(’,’) + 1(\0) = 72 . The length of
each component is 11 since it is a 32 bit signed number.
In gpsinfo.c strncpy will copy at most 71 characters since
this is the maximum length of the string.

Change-Id: Id24995ed08aa277681d9648054ebddc2c275f6b6