Fix insertion of GPSProcessingMethod tag
GPSProcessingMethod was not being inserted properly because it
was being treated as STRING format which results in "ASCII" being
taken as the processing method. GPSProcessingMethod is defined as
UNDEFINED format with the value containing a character string in
Exif2.2 standard.

1. Set GPSProcessingMethod to UNDEFINED format as defined in Exif2.2
2. Add handling of ASCII value with UNDEFINED format in writeExifTagAndData.
Handling is similar to STRING format, except we have to account for the
'ASCII\0\0\0' header before the string.

b/5448171

Change-Id: I411a67d7d97e35bee66aa582b25f0f6333c0679a
Signed-off-by: Tyler Luu <tluu@ti.com>